Refactor borrow and payment models and services
Removed payment-related fields from the Borrow model and updated related service and repository logic to reflect this change. Adjusted ArchivePaiement to use float for paiement_due. Refactored PaiementService to use PaiementType enums, improved error handling, and updated method signatures for consistency. Cleaned up imports and fixed minor issues in service initialization.

+++ app/repositories/models/archive_paiement.py
@@ -4,7 +4,7 @@ from datetime import date
 
 class ArchivePaiement(Base):
 
-    def __init__(self, id : int | None, id_paiement:int, member:Member,paiement_type:int,paiement_due:int,paiement_date:date):
+    def __init__(self, id : int | None, id_paiement:int, member:Member,paiement_type:int,paiement_due:float,paiement_date:date):
 
     
         super().__init__(id)
